requirements.txt
openai
groq
requests
websockets
python-dotenv
